CI: Fix for migration to pyjwt auth
