+++ ci/praktika/hook_html.py
@@ -156,11 +156,11 @@ class HtmlRunnerHooks:
         print(f"CI Status page url [{report_url_current_sha}]")
 
         if Settings.USE_CUSTOM_GH_AUTH:
-            from praktika.gh_auth_deprecated import GHAuth
+            from .gh_auth import GHAuth
 
             pem = _workflow.get_secret(Settings.SECRET_GH_APP_PEM_KEY).get_value()
             app_id = _workflow.get_secret(Settings.SECRET_GH_APP_ID).get_value()
-            GHAuth.auth(app_key=pem, app_id=app_id)
+            GHAuth.auth(app_id=app_id, app_key=pem)
 
         res2 = not bool(env.PR_NUMBER) or GH.post_pr_comment(
             comment_body=f"Workflow [[{_workflow.name}]({report_url_latest_sha})], commit [{_Environment.get().SHA[:8]}]",
@@ -295,11 +295,11 @@ class HtmlRunnerHooks:
 
         if updated_status:
             if Settings.USE_CUSTOM_GH_AUTH:
-                from praktika.gh_auth_deprecated import GHAuth
+                from .gh_auth import GHAuth
 
                 pem = _workflow.get_secret(Settings.SECRET_GH_APP_PEM_KEY).get_value()
                 app_id = _workflow.get_secret(Settings.SECRET_GH_APP_ID).get_value()
-                GHAuth.auth(app_key=pem, app_id=app_id)
+                GHAuth.auth(app_id=app_id, app_key=pem)
 
             print(f"Update GH commit status [{result.name}]: [{updated_status}]")
             GH.post_commit_status(
