+++ ci/praktika/runner.py
@@ -451,11 +451,11 @@ class Runner:
             workflow.enable_commit_status_on_failure and not result.is_ok()
         ) or job.enable_commit_status:
             if Settings.USE_CUSTOM_GH_AUTH:
-                from .gh_auth_deprecated import GHAuth
+                from .gh_auth import GHAuth
 
                 pem = workflow.get_secret(Settings.SECRET_GH_APP_PEM_KEY).get_value()
                 app_id = workflow.get_secret(Settings.SECRET_GH_APP_ID).get_value()
-                GHAuth.auth(app_key=pem, app_id=app_id)
+                GHAuth.auth(app_id=app_id, app_key=pem)
             if not GH.post_commit_status(
                 name=job.name,
                 status=result.status,
